ran flutter clean -> flutter pub get -> flutter pub upgrade and reauthentication now works

+++ lib/bizlogic/authentication.dart
@@ -16,7 +16,9 @@ abstract class BaseAuth {
 
   Future<List<String>> fetchSignInMethodsForEmail();
 
-  Future<void> deleteUser();
+  Future<bool> deleteUser();
+
+  Future<String> reauthenticateWithCredential(String email, String password);
 }
 
 class Auth implements BaseAuth {
@@ -60,8 +62,34 @@ class Auth implements BaseAuth {
     return _firebaseAuth.fetchSignInMethodsForEmail(email: user.email.toString());
   }
 
-  Future<void> deleteUser() async{
+  // returns true if the user was successfully deleted
+  // returns false if an error was caught
+  Future<bool> deleteUser() async{
     final FirebaseUser firebaseUser = await _firebaseAuth.currentUser();
-    await firebaseUser.delete();
+    if (firebaseUser != null){
+      try{
+        await firebaseUser.delete();
+      }
+      catch(ERROR_REQUIRES_RECENT_LOGIN){
+        return false;
+      }
+      return true;
+    }
+    else{
+      return false;
+    }
+  }
+
+  // returns null if the user was successfully authenticated 
+  // with the given email and password
+  // if there was an error, returns the error message
+  @override
+  Future<String> reauthenticateWithCredential(String email, String password) async{
+    // TODO: implement reauthenticateWithCredential
+    final FirebaseUser firebaseUser = await _firebaseAuth.currentUser();
+    AuthCredential credential = EmailAuthProvider.getCredential(
+      email: email, password: password);
+    AuthResult result = await firebaseUser.reauthenticateWithCredential(credential);
+    return result.user.uid;
   }
 }
